Moves VERSION determination to Makefile
Adds targets to Makefile to generate:
 - VERSION.json
 - VERSION.txt
 - VERSION.tmp

And removes the corresponding lines from `packaging/build.sh`,
replacing them with appropraite `make` invocations.

The result should be increased autonomy for the Makefile, as
prior to this change, one had to run `packaging/build.sh` before
calling `make`.

Now, the `include VERSION.tmp` line in the Makefile will
cause (GNU) Make to generate VERSION.tmp if it is not present.